Refactored ObjectEqualsGenerator and EquatableEqualsGenerator into own classes
